fix(ci): python-full-coverage — split monitor_handoff.py under the 1500-line limit
monitor_handoff.py grew to 1577 lines this PR, tripping the first-party
line-limit maintainability guard. Extract the three sync-task handoff
functions (_handoff_sync_release_pr_monitor, _complete_release_pr_sync_no_op,
_handoff_sync_feature_pr_monitor) into a new monitor_handoff_sync.py sibling
module (mirroring the existing monitor_handoff_setup.py pattern) and rewire
them through mixins.py. Behavior is unchanged. Retarget the two tests that
monkeypatch make_forge_client in the moved function's namespace
(error_paths parts 015 and 017) to monitor_handoff_sync.

monitor_handoff.py: 1577 -> 996 lines; monitor_handoff_sync.py: 610 lines.
