+++ src/awf/control/executor/mixins.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from awf.control.executor import git_methods as _git_methods
 from awf.control.executor import monitor_handoff as _monitor_handoff
 from awf.control.executor import monitor_handoff_setup as _monitor_handoff_setup
+from awf.control.executor import monitor_handoff_sync as _monitor_handoff_sync
 from awf.control.executor import planning_ops as _planning_ops
 from awf.control.executor import quality_methods as _quality_methods
 from awf.control.executor import state_ops as _state_ops
@@ -55,9 +56,9 @@ class ExecutorDelegatesMixin:
     _dispatch_non_feature_task_kind = _monitor_handoff._dispatch_non_feature_task_kind
     _run_monitor_handoff_profile_setup = _monitor_handoff_setup._run_monitor_handoff_profile_setup
     _build_handoff_pr_monitor = _monitor_handoff._build_handoff_pr_monitor
-    _handoff_sync_release_pr_monitor = _monitor_handoff._handoff_sync_release_pr_monitor
-    _complete_release_pr_sync_no_op = _monitor_handoff._complete_release_pr_sync_no_op
-    _handoff_sync_feature_pr_monitor = _monitor_handoff._handoff_sync_feature_pr_monitor
+    _handoff_sync_release_pr_monitor = _monitor_handoff_sync._handoff_sync_release_pr_monitor
+    _complete_release_pr_sync_no_op = _monitor_handoff_sync._complete_release_pr_sync_no_op
+    _handoff_sync_feature_pr_monitor = _monitor_handoff_sync._handoff_sync_feature_pr_monitor
     _record_monitor_runtime_restart_failed = _monitor_handoff._record_monitor_runtime_restart_failed
 
     _prepare_conformance_salvage_for_execution = (
